issues #41: optimise common Esp.Net usage to a single using statement.
Removed IModelSubject, IEventSubject and IEventPublisher, in reality you'll probably only use and want a single IRouter, or IRouter<TModel> interface
