+++ src/Esp.Net.Tests/BaseModelEventProcessorTests.cs
@@ -17,7 +17,6 @@
 
 using System;
 using System.Collections.Generic;
-using Esp.Net.ModelRouter;
 using NUnit.Framework;
 using Shouldly;
