+++ src/Esp.Net.Tests/HeldEvents/HeldEventTests.cs
@@ -17,7 +17,6 @@
 #if ESP_EXPERIMENTAL
 using System;
 using System.Collections.Generic;
-using Esp.Net.ModelRouter;
 using Esp.Net.Reactive;
 using NUnit.Framework;
 using Shouldly;
