+++ src/Esp.Net.Tests/RouterTests.ModelRouter.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Esp.Net.ModelRouter;
 using NUnit.Framework;
 using Shouldly;
