+++ src/Esp.Net.Tests/RouterTests.ModelSubRouter.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using Esp.Net.ModelRouter;
 using NUnit.Framework;
 using Shouldly;
