+++ src/Esp.Net.Tests/RouterTests.cs
@@ -17,7 +17,6 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
-using Esp.Net.ModelRouter;
 using Esp.Net.Reactive;
 using NUnit.Framework;
 using Shouldly;
