+++ src/Esp.Net/HeldEvents/RouterExt.cs
@@ -19,7 +19,6 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using Esp.Net.Disposables;
-using Esp.Net.ModelRouter;
 using Esp.Net.Reactive;
 using Esp.Net.Utils;
